FinalProject/test/ModelTest/HandFTest.java
import static org.junit.Assert.*;
import org.junit.Before;
import org.junit.Test;

public class HandFTest {

  private HandF hand;
  private CardF card1;
  private CardF card2;

  @Before
  public void setUp() {
    // Cards are initialized correctly according to the CardF validations.
    card1 = new CardF("A", "S"); // Ace of Spades
    card2 = new CardF("K", "H"); // King of Hearts
    // HandF is abstract, so providing a concrete subclass for testing.
    hand = new HandF(card1, card2) {};
  }

  @Test(expected = IllegalArgumentException.class)
  public void testConstructor_NullFirstCard_ThrowsException() {
    new HandF(null, card2) {};
  }

  @Test(expected = IllegalArgumentException.class)
  public void testConstructor_NullSecondCard_ThrowsException() {
    new HandF(card1, null) {};
  }

  @Test
  public void testGetHandScore() {
    // Assuming TexasHoldemHandEvaluator is correctly implemented and returns an expected score
    int score = hand.getHandScore();
    assertTrue("Score should be a positive integer", score >= 0);
    assertEquals("Score should be 30",23, score);
  }

  @Test
  public void testGetHandCombination() {
    String combination = hand.getHandCombination();
    assertNotNull("Combination should not be null", combination);
    assertFalse("Combination should not be empty", combination.isEmpty());
    assertEquals("Combination should be Suited Connector","Connected Cards",combination);
  }

  @Test
  public void testDescribeHand() {
    String description = hand.describeHand();
    assertNotNull("Description should not be null", description);
    assertTrue("Description should include the first card details", description.contains("A"));
    assertTrue("Description should include the second card details", description.contains("K"));
    assertTrue("Description should mention Total Value", description.matches(".*Total Value: \\d+.*"));
    assertTrue("Description should mention Hand Combination", description.contains("Hand Combination"));
  }
}

